LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/dynamic-import/string-compilation-nonce-module-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8800/html/semantics/scripting-1/the-script-element/module/imports-a.js?label=setTimeout because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/html/semantics/scripting-1/the-script-element/module/imports-a.js?label=reflected%20inline%20event%20handlers because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/html/semantics/scripting-1/the-script-element/module/imports-a.js?label=inline%20event%20handlers%20triggered%20via%20UA%20code because it does not appear in the script-src directive of the Content Security Policy.

FAIL setTimeout must inherit the nonce from the triggering script, thus execute promise_test: Unhandled rejection with value: object "TypeError: Importing a module script failed."
PASS direct eval must inherit the nonce from the triggering script, thus execute
PASS indirect eval must inherit the nonce from the triggering script, thus execute
PASS the Function constructor must inherit the nonce from the triggering script, thus execute
PASS reflected inline event handlers must not inherit the nonce from the triggering script, thus fail
PASS inline event handlers triggered via UA code must not inherit the nonce from the triggering script, thus fail
